refactor(entries): extract HistoryEntryComponent into standalone EntryCard
The 293-line HistoryEntryComponent was defined inline in HistorySettings,
making it unavailable to other views. Moving it to src/components/entries/
lets the main Entries feed and any future views share the same card without
importing from a settings module.
